requirements.txt
# pip freeze -r requirements.txt | sed '/freeze/,$ d'
langchain==0.0.274
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4.2.2
tiktoken==0.4.0
loguru==0.7.0
pypdf==3.15.3
embedchain==0.0.41
gradio==3.41.2

